Empty message for EcdsaKoblitz signatures
I noticed an issue with the EcdsaKoblitzSignature2016 codepaths. `_getDataToHash` returns an empty string if options.algorithm is EcdsaKoblitzSignature2016. That means it's currently signing (and verifying) the empty message.

A better fix than this temporary workaround is to add a canonicalizationAlgorithm option (in addition to algorithm). In this case, the expected canonicalization algorithm is LinkedDataSignature2015, so I just added a check for that.